only all github actions to unzip for less falses
only all github actions to unzip for less falses

+++ src/main/java/common/UpdatePlugins.java
@@ -65,7 +65,7 @@ public class UpdatePlugins {
                 }
             }
 
-            if (isZipFile(downloadPath)) {
+            if (isZipFile(downloadPath) && link.toLowerCase().contains("actions") && link.toLowerCase().contains("github")) {
                 extractFirstJarFromZip(downloadPath, outputFilePath);
                 new File(downloadPath).delete();
             } else {
